high-level-application/requirements.txt
asgiref==3.5.1
astroid==2.3.3
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==3.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
click==8.1.3
colorama==0.4.1
cryptography==3.3.1
cycler==0.10.0
datadiff==2.0.0
decorator==4.4.2
Django==4.0.4
entrypoints==0.3
flake8==3.7.9
Flask==2.1.2
Flask-Cors==3.0.10
idna==2.10
importlib-metadata==4.11.3
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==2.1.2
jedi==0.17.2
Jinja2==3.1.2
jupyter-client==6.1.7
jupyter-core==4.6.3
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
MarkupSafe==2.1.1
matplotlib==3.1.2
matplotlib-venn==0.11.5
mccabe==0.6.1
mpmath==1.1.0
munch==2.5.0
numpy==1.17.4
paramiko==2.7.2
parso==0.7.1
pickleshare==0.7.5
prompt-toolkit==3.0.8
pycodestyle==2.5.0
pycparser==2.20
pyepics==3.5.1
PyExecJS==1.5.1
pyflakes==2.1.1
Pygments==2.7.2
pylint==2.4.4
PyNaCl==1.4.0
pyparsing==2.4.5
pyperclip==1.7.0
python-dateutil==2.8.1
python-jsonrpc==0.10.0
pywin32==228
pyzmq==19.0.2
requests==2.25.1
scipy==1.3.3
Send2Trash==1.5.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.2
tornado==6.1
traitlets==5.0.5
tzdata==2022.1
urllib3==1.25.11
wcwidth==0.2.5
Werkzeug==2.1.2
wrapt==1.11.2
zipp==3.8.0
